Import same class and some variables in to the index.js file
