+++ src/class.js
@@ -1,0 +1,7 @@
+export default class Todos {
+  constructor(description, complete, index) {
+    this.description = description;
+    this.complete = complete;
+    this.index = index;
+  }
+}
